requirements.txt
fastapi[standard]==0.117.0
uvicorn==0.37.0
python-dotenv==1.0.0
httpx==0.28.1
python-telegram-bot==22.4
redis==8.2